Move more common things: shard state manager.
