+++ blueflood-dropwizard/src/main/java/com/rackspacecloud/blueflood/dw/ingest/IngestApplication.java
@@ -2,6 +2,8 @@ package com.rackspacecloud.blueflood.dw.ingest;
 
 import com.google.common.base.Joiner;
 import com.rackspacecloud.blueflood.cache.MetadataCache;
+import com.rackspacecloud.blueflood.dw.NotDOAHealthCheck;
+import com.rackspacecloud.blueflood.dw.StateManager;
 import com.rackspacecloud.blueflood.io.IMetricsWriter;
 import com.rackspacecloud.blueflood.service.Configuration;
 import com.rackspacecloud.blueflood.service.CoreConfig;
